rag_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.8.0
python-dotenv==1.0.0
numpy==1.24.3
aiohttp==3.9.0
pydantic==2.5.0
requests==2.31.0